refactor(#427 W4): ComponentNode is deleted — it wrapped a node, and now there is one
`nros::ComponentNode` owned an `nros::Node` member (RFC-0044 Q1 chose wrap over
derive), so a component was NOT a node and every verb had to be forwarded. Its
946 lines move onto `nros::Node`: the `NodeHandle` constructor, the `ok()` latch
and its accessors, `check_declared_depth`, the member-pointer subscription
family, and the ergonomic macros. `component_node.hpp` is gone.

Two things the merge could not do naively, and what measuring them found.

THE COLLISION IS REAL, AND THE TRIGGER IS NOT THE SPELLING WE PREDICTED

`ComponentNode::create_publisher<M>(const char*, const QoS&)` returned by value
and reported failure through the latch; upstream's returns a `shared_ptr` and
throws. As overloads on one type those differ only in SIGNATURE, which C++
resolves silently. The work item predicted `create_publisher<M>("chatter", 10)`
would bind ours, because array-to-pointer decay is a standard conversion where
`std::string` needs a user-defined one. Compiled against the real headers
(gcc 13, `-std=c++17`), it does not:

    create_publisher<M>("chatter", 10)              -> UPSTREAM (shared_ptr)
    create_publisher<M>("chatter", rclcpp::QoS(10)) -> OURS (by value), accepted
                                                       with only a note that ISO
                                                       calls it ambiguous

`int -> size_t` is a standard conversion and rescues upstream's second argument,
so ours is not better in every argument. What DOES collide is the explicit-QoS
spelling — `nros::QoS(int)` is `explicit` and `rclcpp::QoS(size_t)` is not — and
that is the spelling a ported rclcpp file is most likely to carry. A stronger
argument for the rename than the one the item made, not a weaker one.

So the ours-only family takes `_in` NAMES: `create_publisher_in`,
`create_subscription_in`, `create_wall_timer_in`, `create_timer_in`. Third
application of this rule in the campaign, after the reordered C node initialiser
and the clock-taking C timer verb: where a difference would be settled silently
by overload resolution, change the name, not the signature.

THE 192 BYTES WERE 0.3 % OF THE PROBLEM

`sizeof(ComponentNode)` measured 55,784 bytes. The inline
`ParameterServer<256, 8, 4096>` is ~55.5 KB of that, and NO consumer in the tree
uses it (zero `declare_parameter` call sites across all three consuming
packages), so the parameter facade did not move — components inherit `Node`'s
existing hosted store.

Of the two the ruling addressed:

  * the 24-byte error latch STAYS unconditional. It is the error channel a
    `-fno-exceptions` target has instead of a throwing constructor, so putting
    it behind the hosted block would leave the firmware case — the one that
    cannot throw — with no channel at all.
  * the timer pool becomes a template parameter defaulting to zero, realized as
    `template <::size_t MaxTimers> class NodeWithTimers : public Node`.

`Node` itself could not become the template, and that is measured rather than
preferred: `rclcpp::Node` is an alias asserted with `std::is_same`, 218 in-tree
sites spell `Node` with no argument list, and every `Node&` parameter would
otherwise accept exactly one depth — so a component with a pool would not be a
node the executor could take. Deriving keeps the IS-A relation `ComponentNode`
never had while keeping the bytes opt-in.

    sizeof(nros::Node)   200 -> 224, hosted AND freestanding (-nostdinc++,
                         ThreadX shim) — identical, so the capability-layout
                         rule still holds: a probe may gate a METHOD, never
                         `sizeof`.

The member-pointer `create_subscription_in` family is DECLARED in `node.hpp` and
DEFINED in `component.hpp`, because its body calls `nros::bind_subscription` and
`component.hpp` includes `node.hpp` — defining it at the declaration would close
an include cycle.

Deleting `adopt_launch_seed_` removed the tree's last `if constexpr`, so the
C++17 floor no longer has a live justification. Recorded in `nros.hpp` beside the
surviving helper; lowering it is a separate decision (issue 1118 raised it
deliberately) and is not part of W4.

RFC-0047's "one component, several named nodes" does not exist (PR #773
corrected it) — a component owns exactly one node, and what the subnode packages
exercise is several named CALLBACK GROUPS. So this is TWO constructors, not
three.

Sweep: grep -rn --exclude-dir=.git 'ComponentNode\|component_node' .
