+++ packages/api/nros-cpp/include/nros/component.hpp
@@ -102,7 +102,7 @@ inline Result bind_subscription_raw(Node& node, const char* topic, const char* t
 /// C++14 note: `M`, `C`, and `Method` are all template parameters (none is
 /// deducible from a runtime member-pointer argument without storing it). This
 /// mirrors `bind_subscription_raw<C, &C::m>`; the ergonomic
-/// `ComponentNode::create_subscription` member + `NROS_SUBSCRIBE` macro hide the
+/// `Node::create_subscription_in` member + `NROS_SUBSCRIBE` macro hide the
 /// spelling.
 template <typename M, class C, void (C::*Method)(const M& msg)>
 inline Result bind_subscription(Node& node, const char* topic, C* self,
@@ -375,7 +375,262 @@ inline Result bind_action_client(Node& node, ActionClientStorage& storage, Timer
         poll_timer, poll_ms, [](void* ctx) { nros_cpp_action_client_poll(ctx); }, storage.bytes);
 }
 
+// ==== phase-427 W4 — Node's member-pointer subscription family ==============
+//
+// DECLARED in `node.hpp` (on `class Node`), DEFINED here. `node.hpp` cannot
+// hold these bodies: they call `nros::bind_subscription` above, and this file
+// includes `node.hpp`, so a definition there would close an include cycle. The
+// umbrella `nros.hpp` pulls this file in, so the definitions are visible
+// wherever the members are reachable.
+//
+// Both are `_in`-suffixed, and both came off `nros::ComponentNode`. See the
+// rename note on `Node::create_publisher_in` for why a bare `create_subscription`
+// carrying an ours-only signature is the one thing the merge could not ship.
+
+template <typename M, class C, void (C::*Method)(const M& msg)>
+inline void Node::create_subscription_in(const char* topic, const QoS& qos) {
+    if (!this->check_declared_depth(M::TYPE_NAME, topic, qos)) {
+        return;
+    }
+    Result r = ::nros::bind_subscription<M, C, Method>(*this, topic, static_cast<C*>(this), qos);
+    if (!r.ok()) {
+        this->set_error("create_subscription_in", r.raw());
+    }
+}
+
+template <typename M, class C, void (C::*Method)(const M& msg)>
+inline void Node::create_subscription_in(const CallbackGroup& group, const char* topic,
+                                         const QoS& qos) {
+    // phase-403 step 2 — the same boot-time check as the ungrouped form. A
+    // grouped subscription costs the arena exactly what an ungrouped one does,
+    // so leaving this path out would make the declared depth enforceable
+    // everywhere except in the images that use callback groups.
+    if (!this->check_declared_depth(M::TYPE_NAME, topic, qos)) {
+        return;
+    }
+    const nros_cpp_node_t* h = this->ffi_handle();
+    if (h == nullptr) {
+        this->set_error("create_subscription_in", -3);
+        return;
+    }
+    nros_cpp_qos_t ffi_qos = detail::qos_to_ffi(qos);
+    C* self = static_cast<C*>(this);
+    size_t handle = static_cast<size_t>(-1);
+    // phase-402: the group name is a FIELD now, not a trailing argument. Start
+    // from the library's defaults so a future field cannot be left as whatever
+    // this frame happened to hold.
+    nros_cpp_subscription_options_t sub_options = nros_cpp_subscription_default_options();
+    sub_options.callback_group = group.get_name();
+    nros_cpp_ret_t ret = nros_cpp_subscription_register(
+        h, topic, M::TYPE_NAME, "", ffi_qos,
+        [](const uint8_t* data, size_t len, void* ctx) {
+            M msg;
+            if (M::ffi_deserialize(data, len, &msg) != 0) return;
+            (static_cast<C*>(ctx)->*Method)(msg);
+        },
+        self, &handle, &sub_options);
+    if (ret != 0) {
+        this->set_error("create_subscription_in", ret);
+    }
+}
+
+namespace detail {
+
+/// Strips the reference `decltype(*this)` yields, so `&Self::method` is
+/// well-formed inside the ergonomic macros below.
+template <class T> struct strip_ref {
+    using type = T;
+};
+template <class T> struct strip_ref<T&> {
+    using type = T;
+};
+
+/// phase-403 step 2 — the QoS a `NROS_SUBSCRIBE` with no QoS argument gets.
+///
+/// `QoS(N)` is `QoS::default_profile()` with the depth replaced (RELIABLE,
+/// VOLATILE, KEEP_LAST), so filling the declared depth in changes exactly the
+/// one field the declaration spoke about and nothing else.
+///
+/// `DECLARED_DEPTH_UNDECLARED` returns the default profile untouched. That is
+/// the back-compatibility hinge: every call site that existed before this step,
+/// in an image with no declared depth anywhere, gets the same depth-10 profile
+/// it always got, from a `constexpr` branch the optimiser folds away.
+constexpr ::nros::QoS qos_from_declared_depth(int declared) {
+    return (declared == ::nros::DECLARED_DEPTH_UNDECLARED) ? ::nros::QoS::default_profile()
+                                                           : ::nros::QoS(declared);
+}
+
+} // namespace detail
+
 } // namespace nros
+
+// Zephyr's minimal libcpp ships a STUB <new> (guard
+// ZEPHYR_SUBSYS_CPP_INCLUDE_NEW_) that declares nothrow_t but NOT placement
+// new — the NROS_COMPONENT factory's `new (storage) Class(...)` then fails
+// "no matching operator new(sizetype, void*&)" (first hit porting real
+// Autoware components to a Zephyr image; ASI's FVP build used a full-libcpp
+// toolchain and never saw it). Provide the standard non-allocating forms.
+#ifdef ZEPHYR_SUBSYS_CPP_INCLUDE_NEW_
+inline void* operator new(::std::size_t, void* ptr) noexcept {
+    return ptr;
+}
+inline void* operator new[](::std::size_t, void* ptr) noexcept {
+    return ptr;
+}
+inline void operator delete(void*, void*) noexcept {}
+inline void operator delete[](void*, void*) noexcept {}
+#endif
+
+/// Inside a node constructor: subscribe `void Self::method(const Msg&)` to
+/// `topic`. Derives `Self` from `this` so only the message type, method, and
+/// topic are spelled. An optional 4th argument is the QoS.
+///
+/// State the depth on a memory-constrained target. Since phase-403 a
+/// subscription's arena buffer is sized from its own type, so the cost of a
+/// subscription is `(depth + 1) * the type's bound` -- depth is a multiplier on
+/// the largest thing the topic carries, not a count of small slots. Measured on
+/// mr-canhubk344: nine subscriptions at the default depth 10 wanted 86108 bytes
+/// of arena, and the same nine at depth 1 want 24516.
+///
+/// # phase-403 step 2 — the depth is DECLARED, and the two must agree
+///
+/// The system's contract sidecar (`<bringup>/launch/<stem>.contract.yaml`,
+/// `contracts.sub_endpoints.<endpoint>.qos.depth`) is the source of truth for
+/// SIZING, because the arena is compiled before this TU exists. Both authoring
+/// modes are legal:
+///
+///   * **the code states the QoS** — `NROS_SUBSCRIBE(M, m, "/t", nros::QoS(1))`
+///     and `qos: { depth: 1 }`. If the two numbers differ, this macro fails the
+///     BUILD with a `static_assert` naming the topic and both depths.
+///   * **the contract states the QoS** — `NROS_SUBSCRIBE(M, m, "/t")` with
+///     `qos: { depth: 1 }`, and the declared depth fills in.
+///
+/// With NEITHER a declaration nor a QoS the profile is `QoS::default_profile()`
+/// exactly as before, and nothing is asserted: that image has not opted in, and
+/// an image that has not opted in is not an image in error.
+///
+/// # Dispatching on ARGUMENT COUNT
+///
+/// C++17 has no `__VA_OPT__`, so the macro cannot branch on "a QoS was passed"
+/// with the obvious spelling. `_NROS_SUB_PICK` is the standard argument-count
+/// trick instead: the argument list is padded with the handler names, and which
+/// one lands on the `NAME` parameter depends on how many arguments came before
+/// it.
+///
+/// # STATEMENT context
+///
+/// The 4-argument form expands to a `static_assert` FOLLOWED BY the call, so it
+/// is a statement and not an expression: write `NROS_SUBSCRIBE(...);` on its own
+/// line, as every call site in this tree does.
+#define _NROS_SUB_PICK(_1, _2, _3, _4, NAME, ...) NAME
+#define NROS_SUBSCRIBE(...)                                                                        \
+    _NROS_SUB_PICK(__VA_ARGS__, _NROS_SUB_4, _NROS_SUB_3, _NROS_SUB_TOO_FEW, _NROS_SUB_TOO_FEW)    \
+    (__VA_ARGS__)
+
+/// The 3-argument form: no QoS at the call site, so the DECLARED depth supplies
+/// one. Nothing to assert — there is only ever one number.
+#define _NROS_SUB_3(Msg, method, topic)                                                            \
+    this->template create_subscription_in<                                                         \
+        Msg, ::nros::detail::strip_ref<decltype(*this)>::type,                                     \
+        &::nros::detail::strip_ref<decltype(*this)>::type::method>(                                \
+        (topic),                                                                                   \
+        ::nros::detail::qos_from_declared_depth(::nros::declared_depth(Msg::TYPE_NAME, (topic))))
+
+/// The 4-argument form: the call site states a QoS, so the two numbers must
+/// agree and the BUILD is where that is settled. `#topic` is the topic as the
+/// call site wrote it — the only way a string can reach a `static_assert`
+/// message in C++17.
+#define _NROS_SUB_4(Msg, method, topic, qos)                                                       \
+    NROS_ASSERT_DECLARED_DEPTH(Msg::TYPE_NAME, (topic), (qos), #topic);                            \
+    this->template create_subscription_in<                                                         \
+        Msg, ::nros::detail::strip_ref<decltype(*this)>::type,                                     \
+        &::nros::detail::strip_ref<decltype(*this)>::type::method>((topic), (qos))
+
+/// Fewer than three arguments. Named rather than left undefined so the error is
+/// about the CALL and not about an identifier the reader has never seen.
+#define _NROS_SUB_TOO_FEW(...)                                                                     \
+    static_assert(false, "NROS_SUBSCRIBE takes (Msg, method, topic) or "                           \
+                         "(Msg, method, topic, qos) -- with the QoS omitted, the depth "           \
+                         "declared in the contract sidecar supplies it.")
+
+/// A subscription whose TOPIC is not a compile-time constant.
+///
+/// The compile-time check needs the topic as a constant expression to key the
+/// table with; a topic built at runtime or forwarded through a variable has
+/// none, so a call site like that names itself here and takes the BOOT-TIME
+/// check in `Node::check_declared_depth` instead. That check is the same
+/// comparison against the same table, and it halts boot naming the topic and
+/// both depths.
+///
+/// This is the fallback and not the default on purpose: a build failure is
+/// cheaper than a boot failure, and making the dynamic spelling explicit keeps
+/// the count of call sites that gave up the compile-time check visible.
+#define NROS_SUBSCRIBE_DYNAMIC(Msg, method, topic, qos)                                            \
+    this->template create_subscription_in<                                                         \
+        Msg, ::nros::detail::strip_ref<decltype(*this)>::type,                                     \
+        &::nros::detail::strip_ref<decltype(*this)>::type::method>((topic), (qos))
+
+/// Inside a `nros::NodeWithTimers<N>` constructor: create a repeating timer
+/// calling `void Self::method()` every `period_ms`. Derives `Self` from `this`.
+///
+/// The verb is `create_wall_timer_in` (phase-427 W4): the pool-parked form takes
+/// no storage argument, which would otherwise differ from upstream's
+/// `create_wall_timer(duration, callback)` by SIGNATURE alone.
+#define NROS_CREATE_WALL_TIMER(period_ms, method)                                                  \
+    this->template create_wall_timer_in<                                                           \
+        ::nros::detail::strip_ref<decltype(*this)>::type,                                          \
+        &::nros::detail::strip_ref<decltype(*this)>::type::method>((period_ms))
+
+// -- NROS_COMPONENT(Class) ---------------------------------------------------
+//
+// Marks a `nros::Node`-derived class as the pkg's rclcpp-faithful (IS-A-node)
+// component. Parallels `NROS_NODE_REGISTER` (node_pkg.hpp), but for the
+// construct-with-handle ctor shape.
+//
+// Emits the **factory** (placement-new with the entry's executor node handle) +
+// the qualified **class** string + a **shape:"rclcpp"** marker. There is NO
+// `sizeof`/`alignof` metadata: the typed codegen entry `#include`s the component
+// header, so `sizeof(Class)` / a `Storage<Class>` is a compile-time fact there —
+// not a codegen input.
+//
+// `NROS_PKG_NAME` is the cmake-injected (pre-sanitised) pkg token, same source
+// as `NROS_NODE_REGISTER`. Hand-written pkgs `#define NROS_PKG_NAME my_pkg`
+// before including this header.
+
+#ifndef NROS_PKG_NAME
+#define NROS_PKG_NAME unknown
+#endif
+
+#define _NROS_COMP_CAT_(a, b) a##b
+#define _NROS_COMP_CAT(a, b) _NROS_COMP_CAT_(a, b)
+#define _NROS_COMP_STR_(x) #x
+#define _NROS_COMP_STR(x) _NROS_COMP_STR_(x)
+
+#define _NROS_COMP_FACTORY_SYM(pkg) _NROS_COMP_CAT(__nros_component_factory_, pkg)
+#define _NROS_COMP_CLASS_SYM(pkg) _NROS_COMP_CAT(__nros_component_class_, pkg)
+#define _NROS_COMP_SHAPE_SYM(pkg) _NROS_COMP_CAT(__nros_component_shape_, pkg)
+
+/// Register a `nros::Node`-derived class as the pkg's component. Emits:
+///  - `__nros_component_factory_<pkg>(void* storage, void* node_handle)` — a
+///    C-ABI factory that placement-news `Class(nros::NodeHandle(node_handle))`
+///    into the entry-owned arena slot and returns it as `nros::Node*`.
+///  - `__nros_component_class_<pkg>` — the `"<pkg>::<Class>"` string for lint.
+///  - `__nros_component_shape_<pkg>` — the `"rclcpp"` shape marker.
+///
+/// The derived class MUST have an `explicit Class(nros::NodeHandle)` ctor (it
+/// forwards the handle + the node name to the `Node` base).
+///
+/// The factory returns `::nros::Node*` — phase-427 W4. It used to return
+/// `::nros::ComponentNode*`, a type that WRAPPED a node; the merged type IS one,
+/// so the entry's post-construct `ok()` check now reads the node itself.
+#define NROS_COMPONENT(Class)                                                                      \
+    extern "C" ::nros::Node* _NROS_COMP_FACTORY_SYM(NROS_PKG_NAME)(void* storage,                  \
+                                                                   void* node_handle) {            \
+        return new (storage) Class(::nros::NodeHandle(node_handle));                               \
+    }                                                                                              \
+    extern "C" const char _NROS_COMP_CLASS_SYM(NROS_PKG_NAME)[] =                                  \
+        _NROS_COMP_STR(NROS_PKG_NAME) "::" _NROS_COMP_STR(Class);                                  \
+    extern "C" const char _NROS_COMP_SHAPE_SYM(NROS_PKG_NAME)[] = "rclcpp"
 
 /// Convenience: bind a component subscription member without spelling the
 /// template arguments. `Msg` is unused at runtime (the raw path is type-erased on
